Move bookmark button component to single post
